Split of classes into separate files
